+++ sdk/schemaregistry/azure-schemaregistry-avroencoder/samples/sync_samples/encode_and_decode_with_message_content.py
@@ -84,7 +84,7 @@ if __name__ == "__main__":
         credential=token_credential,
     )
     encoder = AvroEncoder(
-        client=schema_registry, group_name=GROUP_NAME, auto_register_schemas=True
+        client=schema_registry, group_name=GROUP_NAME, auto_register=True
     )
     event_data = encode_message_content_dict(encoder)
     decoded_content = decode_with_content_and_content_type(encoder, event_data)